Fix Startup.cs IConfigService renamed to be similar to IDeviceService
