+++ src/AzureIoTHub.Portal/Server/Services/IConfigServices.cs
@@ -1,16 +1,15 @@
 ﻿// Copyright (c) CGI France. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace AzureIoTHub.Portal.Server.Interfaces
-{
-    using System.Collections.Generic;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.Devices;
+using Microsoft.Azure.Devices;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
-    public interface IConfigs
+namespace AzureIoTHub.Portal.Server.Services
+{
+    public interface IConfigService
     {
         Task<IEnumerable<Configuration>> GetAllConfigs();
-
         Task<Configuration> GetConfigItem(string id);
     }
 }
